move CopyButton to new composable
