+++ packages/frontend-shared/src/gql-components/CopyButton.vue
